(refactor) Result 관련 domain refactoring
