+++ src/main/java/com/grepp/moodlink/app/model/result/dto/BookDto.java
@@ -1,11 +1,12 @@
-package com.grepp.moodlink.infra.app.model.result.dto;
+package com.grepp.moodlink.app.model.result.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 
 @Data
 @AllArgsConstructor
-public class SongDto {
+public class BookDto {
     private String name;
     private String imgUrl;
     private String externalLink;
