+++ src/main/java/com/grepp/moodlink/app/model/result/dto/MovieDto.java
@@ -1,12 +1,11 @@
-package com.grepp.moodlink.infra.app.model.result.dto;
+package com.grepp.moodlink.app.model.result.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 
 @Data
 @AllArgsConstructor
-public class BookDto {
+public class MovieDto {
     private String name;
     private String imgUrl;
     private String externalLink;
