+++ src/main/java/com/grepp/moodlink/app/model/result/dto/SongDto.java
@@ -1,11 +1,11 @@
-package com.grepp.moodlink.infra.app.model.result.dto;
+package com.grepp.moodlink.app.model.result.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class MovieDto {
+public class SongDto {
     private String name;
     private String imgUrl;
     private String externalLink;
