--- src/main/java/com/grepp/moodlink/infra/app/controller/web/main/MainController.java
@@ -1,20 +1,0 @@
-package com.grepp.moodlink.infra.app.controller.web.main;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-
-@Controller
-@Slf4j
-@RequestMapping("main")
-public class MainController {
-
-    @GetMapping
-    public String signin(Model model){
-
-        return "main/main";
-    }
-}
