--- src/main/resources/templates/main/main.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-
-<h1>임시 메인화면</h1>
-</body>
-</html>
