--- src/main/resources/templates/member/mypage.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-
-<h1>임시 유저화면(마이페이지)</h1>
-</body>
-</html>
