--- src/main/resources/templates/recommend/recommend.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-
-<h1>임시 추천화면</h1>
-</body>
-</html>
