+++ src/main/resources/templates/result/result.html
@@ -96,12 +96,12 @@
 </main>
 
   <hr>
-  <form action="/main" method="get">
+  <form action="/signin" method="get">
     <button type="submit">메인화면(헤더로 빼야됨, 모든 Role에 대해 적용)</button>
   </form>
   <br>
   <br>
-  <form action="/member/signin" method="get">
+  <form action="/signin" method="get">
     <button type="submit">로그인화면으로 이동 및 로그아웃 (헤더로 빼야됨, User일 때 Logout, anonymous일 때 Login)</button>
   </form>
   <br>
@@ -111,7 +111,7 @@
   <hr>
 
 
-  <form action="/member/mypage" method="get">
+  <form action="/signin" method="get">
     <button type="submit">마이페이지(User일 때 적용, anonymous의 경우 클릭 시 "회원만 가능한 기능입니다" 출력 후 disable)</button>
   </form>
   <br>
